requirements.txt
openai
websockets
PyAudio
python-dotenv
pydantic
asyncio
numpy